requirements.txt
# Sozialrecht Neo4j RAG System - Dependencies

# Core Dependencies
neo4j==5.25.0
sentence-transformers==3.3.1
numpy>=1.26.0,<2.0.0
langchain==0.3.7
langchain-core==0.3.15

# PDF Processing with Docling
docling==2.14.0
docling-core>=2.12.1,<3.0.0
pdfplumber==0.11.4

# CLI and Utilities
tqdm==4.67.1
tabulate==0.9.0
python-dotenv==1.0.1
pyyaml==6.0.2

# XML Processing and neo4j-graphrag
lxml==5.3.0
requests==2.32.3

# neo4j-graphrag for knowledge graph construction
neo4j-graphrag==0.9.3

# Optional: Azure integration (if using cloud)
# azure-keyvault==4.2.0
# azure-identity==1.19.0
